BO: Fix specific price using combinations
