+++ src/PrestaShopBundle/Controller/Admin/CombinationController.php
@@ -4,8 +4,10 @@ namespace PrestaShopBundle\Controller\Admin;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use PrestaShop\PrestaShop\Adapter\CombinationDataProvider;
 use PrestaShopBundle\Form\Admin\Product\ProductCombination;
+use PrestaShopBundle\Model\Product\AdminModelAdapter as ProductAdminModelAdapter;
 
 class CombinationController extends Controller
 {
@@ -37,5 +39,46 @@ class CombinationController extends Controller
         }
 
         return $response->create($result);
+    }
+
+    /**
+     * get All Combinations for a product
+     *
+     * @param int $idProduct The product id
+     *
+     * @return string Json
+     */
+    public function getProductCombinationsAction($idProduct)
+    {
+        $response = new JsonResponse();
+
+        //get product
+        $productAdapter = $this->container->get('prestashop.adapter.data_provider.product');
+        $product = $productAdapter->getProduct((int)$idProduct);
+
+        //get combinations
+        $modelMapper = new ProductAdminModelAdapter(
+            $product,
+            $this->container->get('prestashop.adapter.legacy.context'),
+            $this->container->get('prestashop.adapter.admin.wrapper.product'),
+            $this->container->get('prestashop.adapter.tools'),
+            $this->container->get('prestashop.adapter.data_provider.product'),
+            $this->container->get('prestashop.adapter.data_provider.supplier'),
+            $this->container->get('prestashop.adapter.data_provider.warehouse'),
+            $this->container->get('prestashop.adapter.data_provider.feature'),
+            $this->container->get('prestashop.adapter.data_provider.pack'),
+            $this->container->get('prestashop.adapter.shop.context')
+        );
+
+        $combinations = $modelMapper->getAttributesResume();
+
+        $combinationList = array();
+        foreach ($combinations as $combination) {
+            $combinationList[] = ['id' => $combination['id_product_attribute'], 'name' => $combination['attribute_designation']];
+        }
+
+        $response->setData($combinationList);
+
+        return $response;
     }
 }
