+++ src/PrestaShopBundle/Resources/config/admin/routing_combination.yml
@@ -6,3 +6,13 @@ admin_combination_generate_form:
     defaults:
         _controller: PrestaShopBundle\Controller\Admin\CombinationController::generateCombinationFormAction
         combinationIds: 0
+
+admin_get_product_combinations:
+    path:     /product-combinations/{idProduct}
+    methods:  [GET]
+    defaults:
+        _controller: PrestaShopBundle\Controller\Admin\CombinationController::getProductCombinationsAction
+        _format: json
+        idProduct: 0
+    requirements:
+        idProduct: \d+
